requirements.txt
plaid-python
python-dotenv
requests
flask
boto3
bcrypt
psycopg2-binary
alembic
